refactor callbackHandler
